requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
cachetools==5.5.1
celery==5.4.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==5.1.6
django-auditlog==3.0.0
django-cors-headers==4.7.0
django-debug-toolbar==5.0.1
django-ranged-response==0.2.0
django-rql==4.4.1
django-simple-captcha==0.6.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-yasg==1.21.10
flake8==7.1.2
gunicorn==20.1.0
inflection==0.5.1
kombu==5.5.1
lark-parser==0.11.0
lib-rql==2.0.2
mccabe==0.7.0
packaging==24.2
pika==1.3.2
pillow==11.1.0
prompt_toolkit==3.0.50
psycopg2==2.9.10
psycopg2-binary==2.9.10
pycodestyle==2.12.1
pyflakes==3.2.0
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.2
setuptools==78.1.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.1
uritemplate==4.1.1
vine==5.1.0
wcwidth==0.2.13
